refactor: separate static merging logic
